feat: moneyd local and btp open signup server
this adds the CLI command 'moneyd local' to run a connector that exposes a BTP server on port 7768 and forwards packets between connected accounts
it also fixes the BTP server setup. the BtpService is bidirectional so it needs to be added as an OutgoingService to the end of the service chain. however, it also needs to take an IncomingService to handle incoming packets. this changes create_server to return an outgoing-only service that can be put into, for example, the RouterService. then, you can call btp_service.handle_incoming(...) with the RouterService, which transforms the btp_service into a bidirectional BtpService that implements both IncomingService and OutgoingService