fix(btp): don't close the WS connections too early or let them hang
before, i thought that we should close the WS connections when the last clone of a BtpService is dropped. however, this causes incoming servers listening on the connections to disconnect immediately. on the other hand, if we don't make sure to close the websockets the SPSP::pay function hangs because the sockets stay open.

this adds an explicit close method to the BtpService. it doesn't seem like an optimal solution but it does allow the connections to be explicitly closed when they should be, while leaving them open by default so that handling incoming packets works properly.