fix(stream): shared secret generation
fixes and simplifies the way the ConnectionGenerator and StreamReceiverService generate and regenerate the shared secret.
this removes the unnecessary step of including the original url in encrypted in the ILP address.
the connection generator now uses a method that allows the receiver service to easily check whether a packet is for it and pass it on to the next outgoing service if not.
this is necessary to allow the receiver service to be used in a pipeline of other outgoing services, such as in the interledger-node