Settlement Engine first approach (#101)
* feat: Add expiry shortener service to reduce the expiry time of packets before 

* feat(redis): encrypt / hmac tokens before storing to db

* feat: Add rate limiting service (packet based / money throughput) - needs redis-cell

* refactor(service): include original_amount in OutgoingRequest

this allows us to decouple the exchange rate and balance services (before, they needed to be together because the balance service needs access to the incoming and outgoing amounts but only the exchange rate service had access to both)
this also makes it easier to implement an outgoing service that tracks packet statistics or saves them to some persistence layer, because such a service will also want access to the incoming and outgoing amounts
note that the field is called "original_amount" only because it seemed strange to add an "incoming_amount" for the services that create OutgoingRequests on their own (only CCP for now but this would also apply to a multi-path STREAM sender that directly creates OutgoingRequests)

* refactor(balances): split balance into payable, receivable, pending {incoming, outgoing}

as described in https://forum.interledger.org/t/what-should-positive-negative-balances-represent/501/5\?u\=emschwartz
adds readme to describe the redis store
removes the pre-funding account functionality (accounts should be created, then funded, and deleted after some time if they are unused)
adds example redis.conf file

* refactor(redis): track single balance and prepaid_amount

based on https://forum.interledger.org/t/what-should-positive-negative-balances-represent/501/26\?u\=emschwartz

* ci: use latest redis version

because version 3.2 doesn't support redis modules

* test: 3 node end-to-end test

including 2 peers using HTTP with one having a BTP client connected
sends SPSP payments in both directions to test connectivity

* feat(node): connect BTP client on startup

still need to make sure this reconnects and connects when new accounts are added

* refactor(http): don't assume client service is the last one

the error message was confusing before because it was http-specific rather than just saying that no route was found

* feat(btp): ping + pong

* fix(node): listen for incoming packets on BTP client

* refactor(node): http before btp in outgoing pipeline

* chore: better logging in btp, http, and spsp

* feat(ccp): request routes on startup

send a RouteControlRequest to all accounts that we are willing to receive routes from when the CCP server starts up
this fixes an issue in the three_nodes test. before, we wanted to wait to connect the BTP client to make sure the BTP server was listening. however, the node with the BTP server needed to send a route broadcast to the client node for it to know how to deliver the SPSP payment, but the client wasn't yet connected and the route broadcast failed. now, the client connects and immediately requests routes from the server

* feat: remove xrp details from node

the settlement method-specific details will be communicated through a different means as suggested in https://forum.interledger.org/t/ledger-adapter-api/553/2\?u\=emschwartz

* feat(node): [WIP] insert admin account by default

instead of requiring that account 0 be inserted before the node is run
also uses the `config` crate to enable configuring the node via config file and env vars instead of CLI arguments

* feat: separate admin auth from default account

admin auth details are passed into the node constructor
CCP server sets the to/from on the request to the same account instead of sending from the default account
default spsp account is configured separately
this resolves issues around race conditions with needing to insert the "default account" first

* chore: remove settlement engine

will be replaced by an external settlement adapter, as discussed in https://forum.interledger.org/t/settlement-architecture/545

* feat(node): default values for redis, http and btp addresses

* feat(settlement): Implement a first approach for the Settlement Engine connection

* Rename Generics for consistency (#99)

* improvement(balance_service) Propagate packet to next peer instead of waiting for store operation to finish (#96)

* Implement TryFrom trait where applicable (#95)

* Disable coverage reports due to https://github.com/emschwartz/interledger-rs/issues/78