refactor(engines/api/store): refactor idempotency ouf of the engines and enforce it at the API level
The exposed API is always idempotent. We avoid redundant code by
wrapping the engine calls in an indempotency check inside a closure.

This allows an engine implementer to not worry about any idempotency
details, since it's already implemented.

IdempotentEngineStore is implemented for Redis as well. There is some code duplication because the Idempotent Engine trait is implemented once for the Redis of the Engine and once for the Redis of the Store crate, but this allows avoiding having to import the full redis store inside the settlement engines crate