fix(service, examples): reject invalid usernames during account creation
The new username feature doesn't allow names to contain special
characters other than underscore, but the simple example had usernames
of the form "node-a". However, this was only being detected during
account lookup, whereas account creation was bypassing the validation
check. This occurred because the intended function by which to construct
a valid instance of Username, `Username::from_str`, was being explicitly
invoked during account lookup; for account creation, the instance of
Username was constructed directly via the Deserialize implementation
generated by Serde. Serde has an attribute hook that allows types to run
custom validation logic via the TryFrom trait; we implement TryFrom for
Username here in terms of the existing FromStr impl (in some future
version of Rust this may be unnecessary, as specialization should allow
TryFrom and FromStr to be defined in terms of each other within libstd).