fix(api): Do not expose balance/prepaid_amount
Updating an account is done by creating a new account with the provided account details and overwriting the existing account with them. When initializing a new account, we initialize balance/prepaid_amount to 0. In order to avoid resetting account balances on account update, we get the account's balances before overwriting it (this does not have extra overhead, since we were doing a check if the account exists already)