chore(api): box warp filters to speed up compilation
theoretically this will make the filters slightly slower at runtime. if the performance hit becomes a problem we can make it so the Filters are only boxed in debug builds. also, most of the API endpoints are for management and not hit in the flow of sending ILP packets so the performance is slightly less important