fix(ccp): send own address in update for epoch 0
I'm not entirely sure this behavior is correct.
However, we now can't guarantee that there is an account with the address of the CCP server (and there probably isn't), so the node isn't sending its own address.
This ensures the node sends its own address at least in the first update.
Right now it won't send its address again if the address changes though.