perf(router): don't clone the routing table HashMap on each packet
the routing table was being cloned for every packet, including the account IDs (UUIDs) for every account. this changes it to return an Arc to avoid copying the underlying data.
this also changes the type of the address prefix to be strings instead of Bytes, because they must be valid UTF8 strings and there were lots of places in the code where we needed to handle parsing the bytes to utf8 where it couldn't actually fail.