feat(settlement): add settlement-core module
This module defines all the functionalities necessary to implement
1. Idempotent APIs
2. APIs which handle precision loss

It also provides default implementations for the above with Redis as a backend

In particular, the  can be called with any engine that conforms to the RFC's trait, and it will return a Warp filter which can be used to run a settlement engine.

This allows an engine implementer to not have to worry about any of the idempotency/precision loss or API related features, and lets them focus on implementing the core business logic of their engine.