test(two_nodes_btp): work-around for inserting account atomicity problem #464
The problemn is that `ilp_address` and the ILP address of an account that is being inserted are not
treated as fully atomic. It should be fixed later after SQL store is integrated. This fix is not an
essential solution and just a temporary work-around, not to let the test fail.

Signed-off-by: Taiga Nakayama <dora@dora-gt.jp>