feat: SQLite database backend, foundation
This commit adds the ability to run a server backed by an SqliteStore as
opposed to a RedisStore, although currently all the necessary Store
trait methods are doing the least amount of work necessary and all the
_un_necessary Store trait methods are simply stubbed out with
`unimplemented!()`. The result is that the running node is able to
successfully respond to the status endpoint, "GET /".