ci: remove deprecated `--all` command from `cargo test`
It's been deprecated in favor of `--workspace`, to distinguish from
`--all-targets` and `--all-features`.

Meanwhile, I've removed the `--all` flag from the `cargo fmt` command,
because `cargo fmt`'s default behavior is already to format all crates
in the workspace. Instead, the `--all` flag tells it to also format
locally-vendored or path-based dependencies in the workspace, which we
neither have nor want.