fix(ccp): fix Mutex errors
Write locks should not be in the same scope as await'ed futures.
We fix this by scoping the write lock for as long as needed, and
then yield only the result of the computation. When the scope is over,
the lock gets dropped, so we can safely use the future.