Improve Error handling (#613)
The objective of this PR is to solidify the errors across the codebase (while now they're just ()). In addition, we'd like to reduce the usage of map_err, since it can get annoying to have to do that everywhere.

* feat: add ilp-errors crate

This is consumed by most crates and is very low in the dependency chain. I decided to create this due to wanting to implement From<RedisError> for T, From<T> for warp::Reject and From<T> for ApiError, where T is each error type.

Had I not introduced a separate crate, I'd have to add a new module for errors in each crate and add warp as a direct dependency to them which I preferred to avoid, and keep it as an implicit dependency via the errors crate.

The error enums are non_exhaustive so that they're future proofed

* feat(service): Add error to Address and Account store

* test(router): fix tests for new error types

* feat(http): Add error to Http store

* feat(btp): Add error to BTP store and expand client errorss

* test(stream): fix tests for new error types

* feat(ccp): Add error to CCP store and simplify return types of helpers

* feat(settlement): add errors to the stores

* feat(service-util): add errors to the stores

* feat(api): remove map_err calls since From<T> for Reject is implemented for the store errors

This means that we no longer need to worry about what error we received from the store
and we can just rely on the From implementation to convert it to the proper
ApiError and then wrap it in a Warp error

* feat(store): add hard errors to the store and rely on the From<RedisError> implementation to remove map_errs

* feat(node): adjust node traits