test: ilp-node configuration
this adds tests for the positive cases and a error precedence cases.
while writing these, it was noticed that the InterledgerNode::default
and the clap::App args are a bit different from each other, meaning that
changing one or the other will be difficult in the future. luckily
however should be caught by these tests, if one forgets the modify the
other place. also, the testing done required that PartialEq bounds
needed to be added.

regarding error precedence: added a few notes on why this exactly
happens. I had assumed that the invalid values would be overwritten
because of the Config::get_str returning a result but that result has a
different meaning than what I had expected.

Signed-off-by: Joonas Koivunen <joonas.koivunen@gmail.com>