refactor(packet): more readable deserialize_envelope
had a lot of issues understanding this, probably because there's a bit
of duplication: if the second argument was `&mut &[u8]` it would be
ready to go as the return value. Also the returned offset is the
starting point for the returned slice. Hopefully the next one coming
over doesn't have to rewrite this a few times until this clarifies.

Signed-off-by: Joonas Koivunen <joonas.koivunen@gmail.com>