doc(readme): remove mention of MSRV 1.45
this had already fallen out of updating with some dependency updates
prompted by cargo audit when it was raised to 1.46 for `if` inside
`const fn`. in #696 the minimum version became 1.51.

since this codebase depends on a lot of the cryptography infrastructure
likely to be updated in the near future due to const generics
availability, it is not worth pursuing any older MSRV.

Signed-off-by: Joonas Koivunen <joonas.koivunen@gmail.com>