refactor(packet): unify handling of trailing bytes
that is, on "strict" reject the inner *and* outer bytes, similar to how
*outer* were already rejected when "strict". for fulfill this changes
how inner trailing bytes were always rejected back to rejecting when
"strict".

refactor out the "outer" bytes check as well, to have the "trailing
bytes" functionality together for more visibility into how this is done
consistently. following the refactor, messages are also unified with
inner/outer.

this follows change from trying to keep the "strict" as adhering as
strict as possible to the RFCs, and perhaps eventually enabling it by
default.

Signed-off-by: Joonas Koivunen <joonas.koivunen@gmail.com>