fix: requests retry count is off by one
Maximum retries were being passed in to the RequestErrorHandler as is.
However in RequestErrorHandler it is defined as maximum attempts.

i.e. `N > max` versus `N == max`

In the future-retries upgrade the check for this was corrected to match
the attempts naming, but this broke the actual usage which assumed it
was retries.

Signed-off-by: Mirko von Leipzig <mirko.vonleipzig@gmail.com>